requirements.txt
aniso8601
annotated-types
anyio
attrs
beautifulsoup4
boto3
botocore
certifi
charset-normalizer
click
exceptiongroup
fastapi
graphene
graphene-sqlalchemy
graphene-sqlalchemy-filter
graphql-core
graphql-relay
greenlet
h11
idna
jmespath
libcst
markdown-it-py
mdurl
mypy-extensions
outcome
promise
pydantic
pydantic_core
Pygments
PyMySQL
pynamodb
PySocks
python-dateutil
python-multipart
pytz
PyYAML
requests
rich
Rx
s3transfer
selenium
singledispatch
six
sniffio
sortedcontainers
soupsieve
SQLAlchemy
starlette
starlette-graphene3
strawberry-graphql
trio
trio-websocket
typer
typing-inspect
typing_extensions
urllib3
uvicorn
wsproto
